added initial outline FeedbackDatasetSharedBase
